pyproject.toml
[tool.poetry]
name = "llm-agent-gateway"
version = "0.1.0"
description = "LLM Agent Gateway"
authors = ["Your Name <you@example.com>"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
httpx = "^0.26.0"
python-dotenv = "^1.0.0"
structlog = "^23.2.0"
openai = "^1.6.1"
google-generativeai = "^0.3.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.9"
mypy = "^1.8.0"
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
